// src/app/models/view-presence.ts
// Copyright 2021 Carnegie Mellon University. All Rights Reserved.
// Released under a MIT (SEI)-style license. See LICENSE.md in the project root for license information.

export interface ViewPresence {
  id: string;
  userId: string;
  userName: string;
  viewId: string;
  online: boolean;
  teamIds: string[];
}